feature/improve-types
